Deleted reflex; Removed outdated comments
Removed commented html from home.php that had been moved to about.php.
Deleted reflex.js and related link in index.php.

Reflex won't be used in the website since it disables responsive
elements. There is no need to have it involved in our project any
longer. If, in the future, we decide otherwise, then it can be
re-downloaded and installed in three minutes.